requirements.txt
pymongo==4.3.3
kafka-python==2.0.2